Implemented the field class
The field class was implemented and the interface was slightly
changed(The fill method was renamed to occupy). This change also
introduces a change in the interface requirements for the Grid class.
Now the Grid class has to be constructible like this:

    Grid{width, height, field}

where width and height are of type Grid::size_type and field of type
Grid::value_type